requirements.txt
pandas
mlflow
notebook
numpy
scikit-learn
matplotlib
python-box
pyyaml
tqdm
ensure
joblib
types-pyyaml
Flask
flask-cors
ipykernel
